requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.38.14
botocore==1.38.14
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.0
colorama==0.4.6
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.3.0
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-htmx==1.21.0
django-resized==1.0.3
django-storages==1.14.6
django-widget-tweaks==1.5.0
idna==3.10
jmespath==1.0.1
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
